refactor: change calculate remainCredit
